Create releases from GH workflow (#1507)
- Implement lxd-build GH workflow
- Implement GH workflow for building armbian and raspberry pi images
- Implement release GH workflow

+++ build/buildlib.sh
@@ -173,7 +173,12 @@ function prepare_chroot_raspbian()
   sudo mount -o bind /dev     raspbian_root/dev/
   sudo mount -o bind /dev/pts raspbian_root/dev/pts
 
-  sudo cp /usr/bin/qemu-aarch64-static raspbian_root/usr/bin
+  if [[ -f "qemu-aarch64-static" ]]
+  then
+    sudo cp qemu-aarch64-static raspbian_root/usr/bin/
+  else
+    sudo cp /usr/bin/qemu-aarch64-static raspbian_root/usr/bin
+  fi
 
   # Prevent services from auto-starting
   sudo bash -c "echo -e '#!/bin/sh\nexit 101' > raspbian_root/usr/sbin/policy-rc.d"
@@ -302,7 +307,7 @@ function pack_image()
   local IMGNAME="$( basename "$IMG" )"
   echo -e "\n\e[1m[ Pack Image ]\e[0m"
   echo "packing $IMG → $TAR"
-  tar -I pbzip2 -C "$DIR" -cvf "$TAR" "$IMGNAME" && \
+  tar -C "$DIR" -cavf "$TAR" "$IMGNAME" && \
     echo -e "$TAR packed successfully"
 }
